Makefile: Optional REGISTRY

+++ Makefile
@@ -7,13 +7,13 @@ TAG=`jq -r .version package.json`
 
 image:
 	docker build -t ${NAME}:${TAG} .
-	docker tag -f ${NAME}:${TAG} ${REGISTRY}/${NAME}:${TAG}
+	if [ -n "${REGISTRY}" ]; then docker tag -f ${NAME}:${TAG} ${REGISTRY}/${NAME}:${TAG}; fi
 
 tests:
 	npm test
 
 push: image
-	docker push ${REGISTRY}/${NAME}:${TAG}
+	if [ -n "${REGISTRY}" ]; then docker push ${REGISTRY}/${NAME}:${TAG}; fi
 	docker push ${NAME}:${TAG}
-	docker rmi ${REGISTRY}/${NAME}:${TAG}
+	if [ -n "${REGISTRY}" ]; then docker rmi ${REGISTRY}/${NAME}:${TAG}; fi
 	docker rmi ${NAME}:${TAG}
